Revert "Revert "[안여경] sprint4""

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicBinaryContentService.java
@@ -15,7 +15,9 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.*;
+import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
@@ -25,11 +27,14 @@ public class BasicBinaryContentService implements BinaryContentService {
     @Override
     public BinaryContent create(BinaryContentCreateRequest createRequest) {
         //TODO : 에러처리
-        byte[] byteArray = this.imageToByteArray(createRequest.contentFile());
-        BinaryContent binaryContent = new BinaryContent(byteArray);
-        this.binaryContentRepository.save(binaryContent);
+        BinaryContent binaryContent = new BinaryContent(
+                createRequest.fileName(),
+                (long) createRequest.bytes().length,
+                createRequest.contentType(),
+                createRequest.bytes()
+        );
 
-        return binaryContent;
+        return this.binaryContentRepository.save(binaryContent);
     }
 
     //reference : https://www.geeksforgeeks.org/java-program-to-convert-byte-array-to-image/
@@ -77,32 +82,24 @@ public class BasicBinaryContentService implements BinaryContentService {
 
     @Override
     public BinaryContent find(UUID binaryContentId) {
-        BinaryContent binaryContent = this.binaryContentRepository
+        return this.binaryContentRepository
                 .findById(binaryContentId)
                 .orElseThrow(() -> new NoSuchElementException("binaryContent with id " + binaryContentId + " not found"));
-
-        return binaryContent;
     }
 
 
     // Reference : https://www.baeldung.com/java-filter-collection-by-list
     @Override
     public List<BinaryContent> findAllByIdIn(List<UUID> binaryContentIds) {
-        // TODO : 잘 작동하는지 확인할것
-        List<BinaryContent> binaryContents = this.binaryContentRepository.findAll();
-        Set<UUID> binaryContentIdsSet = new HashSet<>(binaryContentIds);
-
-        List<BinaryContent> filteredBinaryContents = binaryContents.stream().filter(binaryContentIdsSet::contains).toList();
-
-        return filteredBinaryContents;
+        return this.binaryContentRepository.findAllByIdIn(binaryContentIds);
     }
 
     @Override
     public void delete(UUID binaryContentId) {
-        BinaryContent binaryContent = this.binaryContentRepository
-                .findById(binaryContentId)
-                .orElseThrow(() -> new NoSuchElementException("binaryContent with id " + binaryContentId + " not found"));
-
+        if (!this.binaryContentRepository.existsById(binaryContentId)) {
+            throw new NoSuchElementException("binaryContent with id " + binaryContentId + " not found");
+        }
+        
         this.binaryContentRepository.deleteById(binaryContentId);
     }
 }
