+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserStatusService.java
@@ -2,9 +2,7 @@ package com.sprint.mission.discodeit.service.basic;
 
 import com.sprint.mission.discodeit.dto.UserStatusCreateRequest;
 import com.sprint.mission.discodeit.dto.UserStatusUpdateRequest;
-import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.entity.UserStatus;
-import com.sprint.mission.discodeit.entity.UserStatusType;
 import com.sprint.mission.discodeit.exception.UserStatusAlreadyExistsException;
 import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.repository.UserStatusRepository;
@@ -25,9 +23,9 @@ public class BasicUserStatusService implements UserStatusService {
     @Override
     public UserStatus create(UserStatusCreateRequest createRequest) {
         //1.  `User`가 존재하지 않으면 예외 발생
-        User user = this.userRepository
-                .findById(createRequest.userId())
-                .orElseThrow(() -> new NoSuchElementException("User with id " + createRequest.userId() + " not found"));
+        if (this.userRepository.existsById(createRequest.userId())) {
+            throw new NoSuchElementException("User with id " + createRequest.userId() + " not found");
+        }
 
         //2.  같은 `User`와 관련된 객체가 이미 존재하면 예외를 발생
         this.userStatusRepository.findById(createRequest.userId()).ifPresent((userStatus) -> {
@@ -38,69 +36,58 @@ public class BasicUserStatusService implements UserStatusService {
         UserStatus userStatus = new UserStatus(createRequest.userId());
 
         //4. DB저장
-        this.userStatusRepository.save(userStatus);
-
-        return userStatus;
+        return this.userStatusRepository.save(userStatus);
     }
 
     @Override
     public UserStatus find(UUID userStatusId) {
-        UserStatus userStatus = this.userStatusRepository
+        return this.userStatusRepository
                 .findById(userStatusId)
                 .orElseThrow(() -> new NoSuchElementException("userStatus with id " + userStatusId + " not found"));
-
-        return userStatus;
     }
 
     @Override
     public List<UserStatus> findAll() {
-        List<UserStatus> userStatuses = this.userStatusRepository.findAll().stream().toList();
-
-        return userStatuses;
+        return this.userStatusRepository.findAll().stream().toList();
     }
 
     @Override
-    public UserStatus update(UserStatusUpdateRequest updateRequest) {
+    public UserStatus update(UUID userStatusId, UserStatusUpdateRequest updateRequest) {
         UserStatus userStatus = this.userStatusRepository
-                .findById(updateRequest.userStatusId())
-                .orElseThrow(() -> new NoSuchElementException("userStatus with id " + updateRequest.userStatusId() + " not found"));
+                .findById(userStatusId)
+                .orElseThrow(() -> new NoSuchElementException("userStatus with id " + userStatusId + " not found"));
 
         userStatus.update(updateRequest.status());
 
         /* 업데이트 후 다시 DB 저장 */
         this.userStatusRepository.save(userStatus);
 
-        UserStatus updatedUserStatus = this.userStatusRepository
-                .findById(updateRequest.userStatusId())
-                .orElseThrow(() -> new NoSuchElementException("userStatus with id " + updateRequest.userStatusId() + " not found"));
-
-        return updatedUserStatus;
+        return this.userStatusRepository
+                .findById(userStatusId)
+                .orElseThrow(() -> new NoSuchElementException("userStatus with id " + userStatusId + " not found"));
     }
 
     @Override
-    public UserStatus updateByUserId(UUID userId, UserStatusType status) {
+    public UserStatus updateByUserId(UUID userId, UserStatusUpdateRequest updateRequest) {
         UserStatus userStatus = this.userStatusRepository
                 .findByUserId(userId)
                 .orElseThrow(() -> new NoSuchElementException("userStatus with userId " + userId + " not found"));
 
-        userStatus.update(status);
+        userStatus.update(updateRequest.status());
 
         /* 업데이트 후 다시 DB 저장 */
         this.userStatusRepository.save(userStatus);
 
-        UserStatus updatedUserStatus = this.userStatusRepository
+        return this.userStatusRepository
                 .findByUserId(userId)
                 .orElseThrow(() -> new NoSuchElementException("userStatus with userId " + userId + " not found"));
-
-        return updatedUserStatus;
     }
 
     @Override
     public void delete(UUID userStatusId) {
-        UserStatus userStatus = this.userStatusRepository
-                .findById(userStatusId)
-                .orElseThrow(() -> new NoSuchElementException("userStatus with id " + userStatusId + " not found"));
-
+        if (!this.userStatusRepository.existsById(userStatusId)) {
+            throw new NoSuchElementException("userStatus with id " + userStatusId + " not found");
+        }
         this.userStatusRepository.deleteById(userStatusId);
     }
 }
